// src/types/types.d.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Query = {
  __typename?: 'Query';
  getUserPayouts: Array<User>;
};


export type QueryGetUserPayoutsArgs = {
  sort: Sort;
};

export enum Sort {
  Ascending = 'ASCENDING',
  Descending = 'DESCENDING'
}

export type User = {
  __typename?: 'User';
  email: Scalars['String'];
  employers?: Maybe<Array<Scalars['String']>>;
  grossPay: Scalars['Float'];
  name: Scalars['String'];
  phoneNumber: Scalars['String'];
};